requirements.txt
cycler==0.10.0
decorator==4.4.0
easydict==1.9
imageio==2.5.0
imgaug==0.4.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.2
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
tensorboardX==1.8
torch==1.5.0
torchvision==0.2.2
torchgeometry==0.1.2
